requirements.txt
psycopg2-binary==2.9.9
python-dotenv==0.19.0
pandas==2.1.2
Faker==19.12.0
sqlalchemy==2.0.22
numpy==1.26.0
